requirements.txt
# Hybrid AI-Quantum Satellite Image Encryption System
# Core dependencies

# Quantum computing
qiskit>=1.0.0
qiskit-aer>=0.13.0

# Deep learning (FlexiMo)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.2
einops>=0.6.1

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.20.0

# Cryptography
cryptography>=41.0.0

# Visualization
matplotlib>=3.7.0

# Utilities
scipy>=1.9.1
